Refactoring and documentation of the API implementation.
The public-facing API has not really changed, but the implementation has
in such a way as to allow for named JS types, such as
`HttpClientRequest` that map to the Java API types.

Documentation is currently generated with each build of the lang-dynjs
module here: https://projectodd.ci.cloudbees.com/view/DynJS/job/dynjs-vertx-module/lastSuccessfulBuild/artifact/target/docs/index.html

The only minor public difference in the refactoring and the old
implementation is that we now have each of the vertx modules hanging off
of the `vertx` object in addition to the properties of those modules
being wired directly to `vertx`. For example:

    // The historical way - still works but is not documented
    // and may eventually be deprecated.
    var vertx = require('vertx');
    var server = vertx.createHttpServer();

    // The new way namespaces things. This is the only change that
    // has been made to the public facing API.
    var vertx = require('vertx');
    var server = vertx.http.createHttpServer();

    // And of course, you can just load the module you care about too
    var http = require('http');
    var server = http.createHttpServer();